fread, reshape2, tryCatch
Learnt more about data.table. Want to use fread because it is so fast and select= is so useful, but for now not use the j indexing, so set option(datatable.fread.datatable=FALSE) and use fread instead of read.table.

Added a version of the summarising using reshape2 melt/cast to learn them.

Added a tryCatch() around the main code so that I don't have to keep setwd() back manually when I make a mistake developing things.

Also puts the fetch.dataset() call back in.